requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
cssselect2==0.7.0
dj-database-url==1.2.0
Django==3.2.18
django-cors-headers==3.10.1
django-environ==0.10.0
django-filter==22.1
django-seed==0.3.1
django-selenium-pdfmaker==0.0.5
xhtml2pdf==0.2.9
djangorestframework==3.14.0
drf-excel==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Faker==17.6.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.2
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.7.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
toposort==1.10
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzlocal==5.0b2
uritools==4.0.1
urllib3==1.26.15
webdriver-manager==3.8.5
webencodings==0.5.1
wsproto==1.2.0
xhtml2pdf==0.2.9
pandas==2.0.0
redis==4.4.2
celery==5.2.7
SQLAlchemy==2.0.9
